Hi Yannick,
I ran converter on every piece of code we have and it is doing fantastic job.

After thoroughly doing it, I was left with few instances that were failing... out of which I threw away ~8 types of issues as they should not be handled by converter at all. They are coding issues that need to be addressed in as3 source. But, I was still left with ~5-6 other cases to be handled by converter. This is , by all probability, FINAL set for converter work :) 

Of these, only 2 need serious work, rest are easy.

Please review these changes. Let me know if you have any questions.

Thanks,
Srikanth